+++ src/leveldb/db/leveldb_main.cc
@@ -3,6 +3,7 @@
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
 #include <stdio.h>
+#include <stdint.h>
 #include "db/dbformat.h"
 #include "db/filename.h"
 #include "db/log_reader.h"
@@ -56,8 +57,8 @@ bool PrintLogContents(Env* env, const std::string& fname,
   Slice record;
   std::string scratch;
   while (reader.ReadRecord(&record, &scratch)) {
-    printf("--- offset %llu; ",
-           static_cast<unsigned long long>(reader.LastRecordOffset()));
+    printf("--- offset %" PRIu64 "; ",
+           static_cast<uint64_t>(reader.LastRecordOffset()));
     (*func)(record);
   }
   delete file;
@@ -92,8 +93,8 @@ static void WriteBatchPrinter(Slice record) {
   }
   WriteBatch batch;
   WriteBatchInternal::SetContents(&batch, record);
-  printf("sequence %llu\n",
-         static_cast<unsigned long long>(WriteBatchInternal::Sequence(&batch)));
+  printf("sequence %" PRIu64 "\n",
+         static_cast<uint64_t>(WriteBatchInternal::Sequence(&batch)));
   WriteBatchItemPrinter batch_item_printer;
   Status s = batch.Iterate(&batch_item_printer);
   if (!s.ok()) {
@@ -165,7 +166,7 @@ bool DumpTable(Env* env, const std::string& fname) {
       }
       printf("'%s' @ %8llu : %s => '%s'\n",
              EscapeString(key.user_key).c_str(),
-             static_cast<unsigned long long>(key.sequence),
+             static_cast<uint64_t>(key.sequence),
              type,
              EscapeString(iter->value()).c_str());
     }
